+++ src/modules/cmdpal/Exts/TemplateExtension/Pages/TemplateExtensionPage.cs
@@ -1,0 +1,30 @@
+﻿// Copyright (c) Microsoft Corporation
+// The Microsoft Corporation licenses this file to you under the MIT license.
+// See the LICENSE file in the project root for more information.
+
+using System;
+using Microsoft.CmdPal.Extensions;
+using Microsoft.CmdPal.Extensions.Helpers;
+
+namespace TemplateExtension;
+
+internal sealed partial class TemplateExtensionPage : ListPage
+{
+    public TemplateExtensionPage()
+    {
+        Icon = new(string.Empty);
+        Name = "TemplateDisplayName";
+    }
+
+    public override ISection[] GetItems()
+    {
+        return [
+            new ListSection()
+            {
+                Items = [
+                    new ListItem(new NoOpCommand()) { Title = "TODO: Implement your extension here" }
+                ],
+            }
+        ];
+    }
+}
